+++ src/WeightItem.java
@@ -1,7 +1,7 @@
 public class WeightItem extends Storeitem
 {
-    protected int weight;
-    public WeightItem(String name, double price, String description, int weight, int itemNum) {
+    protected double weight;
+    public WeightItem(String name, double price, String description, double weight, int itemNum) {
         super(name, price, description, itemNum);
         this.weight = weight;
